fix: change folder structure, add white world map bg
