+++ app/(main)/(home)/(search)/_components/LoadingProgress.tsx
