+++ app/(main)/(home)/(search)/_components/MainHeader.tsx
