+++ app/(main)/(home)/(search)/_components/NavigatorHeader.tsx
