+++ app/(main)/(home)/(search)/_components/help/HelpCard.tsx
