+++ app/(main)/(home)/(search)/_components/history/HistoryCards.tsx
