+++ app/(main)/(home)/(search)/_components/input/LinkCard.tsx
