+++ app/(main)/result/page.tsx
@@ -1,13 +1,11 @@
 'use client';
 
-import AccuracyDashboard from './_components/AccuracyDashboard';
 import FeedbackCard from './_components/FeedbackCard';
 import SummaryCard from './_components/SummaryCard';
 import ArticleCards from './_components/ArticleCards';
 import { useResultStore } from '@/lib/store';
 import { useEffect, useState } from 'react';
 import type { Result } from '@/lib/types';
-import Image from 'next/image';
 import AccuracyThumbnail from './_components/AccuracyThumbnail';
 
 const Result = ({
@@ -37,13 +35,15 @@ const Result = ({
   const accuracy = 100 - analysis_result.fake_possibility;
 
   return (
-    <>
-
+    <div className='p-5 flex flex-col gap-6'>
+      <div className='line-clamp-2 text-4xl font-bold text-white z-10'>
+        {youtube_info.video_title}
+      </div>
       <AccuracyThumbnail accuracy={accuracy} youtube_info={youtube_info} />
       <SummaryCard accuracy={accuracy} summary={analysis_result.summary} />
       <ArticleCards accuracy={accuracy} articles={related_articles} />
       <FeedbackCard accuracy={accuracy} />
-    </>
+    </div>
   );
 };
